Update to bsnes v062r05 release.
To run this, you'll need the DLLs from v062r04's public beta, and the
updated snesreader.dll in the same folder as the WIP. No profiling.

This fixes UPS patching, and it also modifies snesreader to generate
the XML map separately, so that the map can be generated post-
patching.
The new enum classes weren't attaching properly to the config file, so
the input settings, expansion port and region settings are saved
again.
It also converts the S-SMP timers back to unsigned integers instead of
using floating point, no accuracy improvement but much more in line
with hardware.
Lastly, it makes the div register shift left 16 places and pre-shifts
on divide, which is just for aesthetics.

And I'll wait on your tests, FitzRoy. I really hope that Big Run
Jaleco Rally is correct, because I don't have the first idea how to
debug that one. Speedy I can probably handle.

+++ src/smp/ssmp/timing/timing.cpp
@@ -33,7 +33,7 @@ void sSMP::sSMPTimer<frequency>::tick() {
   //stage 0 increment
   stage0_ticks += smp.status.timer_step;
   if(stage0_ticks < frequency) return;
-  stage0_ticks -= (double)frequency;
+  stage0_ticks -= frequency;
 
   //stage 1 increment
   stage1_ticks ^= 1;
